requirements.txt
torch>=2.1.0
torchvision
transformers
peft
accelerate
bitsandbytes
opencv-python
pillow
requests
datasets
scikit-learn
numpy